#11439 Use getLargeUpdateCount if update count exceeds max value
Former-commit-id: 8908bb87837ddccf64708c7bd315251be41b500b

+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/impl/local/LocalStatement.java
@@ -67,7 +67,7 @@ public class LocalStatement extends AbstractStatement<DBCSession>
     }
 
     @Override
-    public int getUpdateRowCount() throws DBCException {
+    public long getUpdateRowCount() throws DBCException {
         return 0;
     }
